move cache tests to the cache namespace
