+++ src/Cache/ClearCacheDirective.php
@@ -8,7 +8,6 @@ use Illuminate\Contracts\Cache\Repository as CacheRepository;
 use Illuminate\Support\Arr;
 use Nuwave\Lighthouse\Schema\Directives\BaseDirective;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
-use Nuwave\Lighthouse\Support\Contracts\CacheKeyAndTags;
 use Nuwave\Lighthouse\Support\Contracts\FieldMiddleware;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 
@@ -20,7 +19,7 @@ class ClearCacheDirective extends BaseDirective implements FieldMiddleware
     protected $cacheRepository;
 
     /**
-     * @var \Nuwave\Lighthouse\Support\Contracts\CacheKeyAndTags
+     * @var \Nuwave\Lighthouse\Cache\CacheKeyAndTags
      */
     protected $cacheKeyAndTags;
